Remove createHoC file and remove injectSheet to withStyles
