+++ packages/react-jss/src/index.js
@@ -3,4 +3,4 @@ export {ThemeProvider, withTheme, createTheming} from 'theming'
 export {default as JssProvider} from './JssProvider'
 export {default as jss} from './jss'
 export {SheetsRegistry, createGenerateId} from 'jss'
-export {default} from './injectSheet'
+export {default} from './withStyles'
